Add trait ExtendCatalogConfig and remove flattened struct
