+++ rust/crd/src/catalog/mod.rs
@@ -1,3 +1,4 @@
+pub mod commons;
 pub mod hive;
 pub mod iceberg;
