+++ rust/operator-binary/src/catalog/iceberg.rs
@@ -1,4 +1,4 @@
-use super::{config::CatalogConfig, FromTrinoCatalogError, ToCatalogConfig};
+use super::{config::CatalogConfig, ExtendCatalogConfig, FromTrinoCatalogError, ToCatalogConfig};
 use async_trait::async_trait;
 use stackable_operator::client::Client;
 use stackable_trino_crd::catalog::iceberg::IcebergConnector;
@@ -13,15 +13,7 @@ impl ToCatalogConfig for IcebergConnector {
         catalog_namespace: Option<String>,
         client: &Client,
     ) -> Result<CatalogConfig, FromTrinoCatalogError> {
-        // Get config that is generated by attributed shared between Hive and Iceberg
-        let mut config = self
-            .common
-            .to_catalog_config(catalog_name, catalog_namespace, client)
-            .await?;
-
-        // As the `HiveAndIcebergCommonAttributes` struct can't know for which connector it is called
-        // we have to overwrite the connector name here
-        config.add_property("connector.name", CONNECTOR_NAME);
+        let mut config = CatalogConfig::new(catalog_name.to_string(), CONNECTOR_NAME);
 
         // No authorization checks are enforced at the catalog level.
         // We don't want the iceberg connector to prevent users from dropping tables.
@@ -29,6 +21,25 @@ impl ToCatalogConfig for IcebergConnector {
         // This decision should be done globally (for all catalogs) by OPA.
         // See https://trino.io/docs/current/connector/iceberg.html
         config.add_property("iceberg.security", "allow-all");
+
+        self.metastore
+            .extend_catalog_config(&mut config, catalog_name, catalog_namespace.clone(), client)
+            .await?;
+
+        if let Some(ref s3) = self.s3 {
+            s3.extend_catalog_config(&mut config, catalog_name, catalog_namespace.clone(), client)
+                .await?;
+        }
+
+        if let Some(ref hdfs) = self.hdfs {
+            hdfs.extend_catalog_config(
+                &mut config,
+                catalog_name,
+                catalog_namespace.clone(),
+                client,
+            )
+            .await?;
+        }
 
         Ok(config)
     }
