+++ rust/operator-binary/src/catalog/mod.rs
@@ -1,3 +1,4 @@
+pub mod commons;
 pub mod config;
 pub mod hive;
 pub mod iceberg;
@@ -26,4 +27,15 @@ pub trait ToCatalogConfig {
         catalog_namespace: Option<String>,
         client: &Client,
     ) -> Result<CatalogConfig, FromTrinoCatalogError>;
+}
+
+#[async_trait]
+pub trait ExtendCatalogConfig {
+    async fn extend_catalog_config(
+        &self,
+        catalog_config: &mut CatalogConfig,
+        catalog_name: &str,
+        catalog_namespace: Option<String>,
+        client: &Client,
+    ) -> Result<(), FromTrinoCatalogError>;
 }
